GH-1: Bingo(コンポーネント）
